requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
Appium-Python-Client==1.0.2
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
idna==2.9
importlib-metadata==1.7.0
more-itertools==8.4.0
ParamUnittest==0.2
pluggy==0.13.1
py==1.9.0
PyMySQL==0.9.3
pytest==4.5.0
PyYAML==5.4.1
requests==2.23.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.5
xlrd==1.2.0
zipp==3.1.0
